+++ chevah/empirical/tests/normal/test_testcase.py
@@ -3,7 +3,7 @@
 """
 Tests for ChevahTestCase.
 """
-from __future__ import with_statement
+import sys
 
 from twisted.internet import defer, reactor
 from twisted.internet.task import Clock
@@ -349,6 +349,37 @@ class TestEmpiricalTestCase(EmpiricalTestCase):
         in a normal testcase.
         """
         raise self.skipTest()
+
+    def test_listenPort(self):
+        """
+        It can be used for listening a dummy connection on a port and address.
+        """
+        address = '127.0.0.1'
+        port = 10000
+
+        with self.listenPort(address, port):
+
+            self.assertIsListening(address, port)
+
+    def test_listenPort_on_loopback_alias(self):
+        """
+        Integration test to check that we can listen on loopback alias.
+
+        This is a system test, but socket operations are light.
+        """
+        if sys.platform.startswith('aix'):
+            # On AIX and probably on other Unixes we can only bind on
+            # existing fixed IP addressed like 127.0.0.1.
+            raise self.skipTest()
+
+        # This is just a test to make sure that the server can listen to
+        # 127.0.0.10 as this IP is used in other tests.
+        address = '127.0.0.10'
+        port = 10070
+
+        with self.listenPort(address, port):
+
+            self.assertIsListening(address, port)
 
 
 class TestEmpiricalTestCaseSkipSetup(EmpiricalTestCase):